rename RSA class and file
i'm working on sorting the code and classes so the name also, so I moved RSA file encrypt functions to rsa_encryptor.py and delete the file_encryptor.py
